nexus/requirements.txt
flask==3.0.3
google-generativeai==0.7.2
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.10.0
gunicorn==22.0.0
Werkzeug==3.0.3
bleach==6.1.0
Pillow==10.3.0
python-dotenv==1.0.1
